Add nullability support (#3)
* Add nullability support for Literal types

Accept list types and dictionary types with optional values. Keeping
the type constraint to non-optional works, but because Any? is
interchangeable to Any type, it causes collection type with optional
values to be double optional type when its elements are accessed.

Make non-literal comparisons to equal to true. Since types other
than what is allowed for PlanOut operator is generalized into
"nonLiteral" (i.e. type-erased), therefore it makes more sense for
non literal types to be true when compared with other non-literal
types despite having different original values.

* Add evaluation state for interpreter

Interpreter's evaluation method is recursive, and there is a way for
inner operators to access Interpreter's getter methods (i.e.
ReadableParams protocol methods) which will trigger another
evaluateExperiment() call.

Previously, any operators that contains any get calls to interpreter
would cause infinite loop because the interpreter would perform yet
another evaluation for the original / initial argument.

This is solved by introducing "evaluating" state, which prevents
the cause of infinite loop.

* Use updateValue in PlanOutOpSimple

Previously, updating values using setter method via subscripting
will remove the key provided that the value is nil. Using the
updateValue method will preserve the nil value instead.

The nil value needs to be preserved, because operators need the
differentiation whether the nil value is due to (1) the key does
not exist, or (2) the key *does* exist but the value is null.

Most operators would throw if the key does not exist, while some
operators (such as Length) allows nil values.

* Fix Condition operator

Previous implementation missed the requirements on processing "cond"
arguments, which contains an array of dictionaries with "if" and
"then" values.

Test cases are also adjusted to the correct behavior.

* Add nullability support for Array and Coalesce ops

Arrays should not compactMap values, and instead return array with
nullable values.

Coalesce will instead perform the compact map as it must return the
first nonnull values from the arrays.

* Get operator throws if var name is not String

* Fix sampling procedure in DefaultSegmentAllocator

The DefaultSegmentAllocator uses Set-backed store, which do not
guarantee the order of the elements contained within the collection.

Since Sample operator depends on the order of the elements within
the collection, it is expected that given the same input and the
same order, it should always produce the same samples.

The fix is done by sorting the Set before it is supplied to the
sampler function.

* Add nullability support for simple operators

* Add nullability support for unary operators

* Add nullability support for binary operators

* Add nullability support for commutative operators

* Add nullability support for comparison operators

* Add nullability support for random operators